Little refactors on various modules
