+++ oda-connectors/http/src/main/resources/config_model.cfg
@@ -1,0 +1,6 @@
+host:localhost
+port:80
+generalPath:commonPathToAllSites
+collectionPath:/odaData
+compressionEnabled:true
+compressionThreshold:0
