+++ oda-datastreams/mqtt/src/test/java/es/amplia/oda/datastreams/mqtt/MqttDatastreamsOrchestratorTest.java
@@ -9,6 +9,7 @@ import es.amplia.oda.core.commons.interfaces.EventPublisher;
 import es.amplia.oda.core.commons.interfaces.Serializer;
 import es.amplia.oda.core.commons.utils.ServiceRegistrationManagerWithKey;
 
+import es.amplia.oda.datastreams.mqtt.configuration.MqttDatastreamsConfiguration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
