--- oda-datastreams/simulator/src/main/java/es/amplia/oda/datastreams/simulator/ConstantDatastreamConfiguration.java
@@ -1,10 +1,0 @@
-package es.amplia.oda.datastreams.simulator;
-
-import lombok.Value;
-
-@Value
-class ConstantDatastreamConfiguration implements SimulatedDatastreamsConfiguration {
-    private String datastreamId;
-    private String deviceId;
-    private Object value;
-}
