--- oda-datastreams/simulator/src/main/java/es/amplia/oda/datastreams/simulator/SimulatedDatastreamsConfiguration.java
@@ -1,4 +1,0 @@
-package es.amplia.oda.datastreams.simulator;
-
-interface SimulatedDatastreamsConfiguration {
-}
