+++ oda-datastreams/simulator/src/main/java/es/amplia/oda/datastreams/simulator/SimulatedDatastreamsManager.java
@@ -2,6 +2,9 @@ package es.amplia.oda.datastreams.simulator;
 
 import es.amplia.oda.core.commons.interfaces.DatastreamsGetter;
 import es.amplia.oda.core.commons.utils.ServiceRegistrationManager;
+import es.amplia.oda.datastreams.simulator.configuration.ConstantDatastreamConfiguration;
+import es.amplia.oda.datastreams.simulator.configuration.RandomDatastreamConfiguration;
+import es.amplia.oda.datastreams.simulator.configuration.SimulatedDatastreamsConfiguration;
 import es.amplia.oda.datastreams.simulator.internal.SimulatedDatastreamsGetterFactory;
 
 import java.util.HashMap;
@@ -9,7 +12,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 
-class SimulatedDatastreamsManager implements AutoCloseable {
+public class SimulatedDatastreamsManager implements AutoCloseable {
 
     private final SimulatedDatastreamsGetterFactory factory;
     private final ServiceRegistrationManager<DatastreamsGetter> registrationManager;
@@ -29,7 +32,7 @@ class SimulatedDatastreamsManager implements AutoCloseable {
         datastreamsGetterCreators.put(RandomDatastreamConfiguration.class, this::createRandomDatastreamsGetter);
     }
 
-    void loadConfiguration(List<SimulatedDatastreamsConfiguration> configuration) {
+    public void loadConfiguration(List<SimulatedDatastreamsConfiguration> configuration) {
         configuration.stream()
                 .map(this::createDatastreamsGetter)
                 .forEach(this::registerDatastreamsGetter);
