+++ oda-datastreams/simulator/src/main/java/es/amplia/oda/datastreams/simulator/configuration/ConstantDatastreamConfiguration.java
@@ -1,0 +1,10 @@
+package es.amplia.oda.datastreams.simulator.configuration;
+
+import lombok.Value;
+
+@Value
+public class ConstantDatastreamConfiguration implements SimulatedDatastreamsConfiguration {
+    private String datastreamId;
+    private String deviceId;
+    private Object value;
+}
