+++ oda-datastreams/simulator/src/main/java/es/amplia/oda/datastreams/simulator/configuration/SimulatedDatastreamsConfiguration.java
@@ -1,0 +1,4 @@
+package es.amplia.oda.datastreams.simulator.configuration;
+
+public interface SimulatedDatastreamsConfiguration {
+}
