+++ oda-datastreams/simulator/src/test/java/es/amplia/oda/datastreams/simulator/ActivatorTest.java
@@ -3,6 +3,7 @@ package es.amplia.oda.datastreams.simulator;
 import es.amplia.oda.core.commons.interfaces.DatastreamsGetter;
 import es.amplia.oda.core.commons.utils.ConfigurableBundleImpl;
 import es.amplia.oda.core.commons.utils.ServiceRegistrationManagerOsgi;
+import es.amplia.oda.datastreams.simulator.configuration.SimulatedDatastreamsConfigurationHandler;
 import es.amplia.oda.datastreams.simulator.internal.SimulatedDatastreamsGetterFactory;
 
 import org.junit.Test;
